Actualizacion Backend v.1

+++ src/main/java/pe/cibertec/inkaproductos/models/UnidadMedida.java
@@ -3,13 +3,10 @@ package pe.cibertec.inkaproductos.models;
 import jakarta.persistence.*;
 import lombok.Data;
 
-@Data
-@Entity
-@Table(name = "unidad_medida")
+@Data @Entity @Table(name = "unidad_medida")
 public class UnidadMedida {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Short uomId;
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer uomId;
     private String codigo;
     private String descripcion;
 }
